Project structure modified
