+++ app/src/main/java/com/github/code/gambit/data/remote/services/file/FileService.kt
@@ -1,4 +1,4 @@
-package com.github.code.gambit.network.api.file
+package com.github.code.gambit.data.remote.services.file
 
 import com.github.code.gambit.data.entity.network.FileNetworkEntity
