+++ app/src/main/java/com/github/code/gambit/data/remote/services/user/UserService.kt
@@ -1,4 +1,4 @@
-package com.github.code.gambit.network.api.user
+package com.github.code.gambit.data.remote.services.user
 
 import com.github.code.gambit.data.entity.network.UserNetworkEntity
