+++ app/src/main/java/com/github/code/gambit/di/NetworkModule.kt
@@ -3,17 +3,17 @@ package com.github.code.gambit.di
 import com.github.code.gambit.data.mapper.network.FileNetworkMapper
 import com.github.code.gambit.data.mapper.network.UrlNetworkMapper
 import com.github.code.gambit.data.mapper.network.UserNetworkMapper
-import com.github.code.gambit.network.api.ApiService
-import com.github.code.gambit.network.api.NetworkDataSource
-import com.github.code.gambit.network.api.NetworkDataSourceImpl
-import com.github.code.gambit.network.api.file.FileService
-import com.github.code.gambit.network.api.file.FileServiceImpl
-import com.github.code.gambit.network.api.url.UrlService
-import com.github.code.gambit.network.api.url.UrlServiceImpl
-import com.github.code.gambit.network.api.user.UserService
-import com.github.code.gambit.network.api.user.UserServiceImpl
+import com.github.code.gambit.data.remote.NetworkDataSource
+import com.github.code.gambit.data.remote.NetworkDataSourceImpl
+import com.github.code.gambit.data.remote.services.ApiService
+import com.github.code.gambit.data.remote.services.file.FileService
+import com.github.code.gambit.data.remote.services.file.FileServiceImpl
+import com.github.code.gambit.data.remote.services.url.UrlService
+import com.github.code.gambit.data.remote.services.url.UrlServiceImpl
+import com.github.code.gambit.data.remote.services.user.UserService
+import com.github.code.gambit.data.remote.services.user.UserServiceImpl
 import com.github.code.gambit.utility.AppConstant
-import com.github.code.gambit.utility.UserManager
+import com.github.code.gambit.utility.sharedpreference.UserManager
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import dagger.Module
