+++ app/src/main/java/com/github/code/gambit/utility/extention/view.kt
@@ -1,4 +1,4 @@
-package com.github.code.gambit.utility
+package com.github.code.gambit.utility.extention
 
 import android.view.View
 import android.view.Window
@@ -7,32 +7,8 @@ import android.view.animation.AnimationUtils
 import android.view.animation.DecelerateInterpolator
 import android.widget.RelativeLayout
 import androidx.core.view.isVisible
-import com.amplifyframework.auth.AuthUserAttribute
-import com.amplifyframework.auth.AuthUserAttributeKey
-import com.amplifyframework.auth.options.AuthSignUpOptions
 import com.github.code.gambit.R
-import com.github.code.gambit.data.model.User
-import com.github.code.gambit.helper.auth.AuthData
 import com.google.android.material.snackbar.Snackbar
-import java.util.Base64
-
-fun String.toBase64(): String {
-    val byteArray = if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
-        Base64.getEncoder().encode(this.toByteArray())
-    } else {
-        android.util.Base64.encode(this.toByteArray(), android.util.Base64.DEFAULT)
-    }
-    return String(byteArray)
-}
-
-fun String.fromBase64(): String {
-    val byteArray = if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
-        Base64.getDecoder().decode(this)
-    } else {
-        android.util.Base64.decode(this, android.util.Base64.DEFAULT)
-    }
-    return String(byteArray)
-}
 
 fun Window.setStatusColor(color: Int) {
     this.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
@@ -84,8 +60,7 @@ fun List<View>.showAll() {
 }
 
 fun View.snackbar(message: String, view: View) {
-    Snackbar
-        .make(this, message, Snackbar.LENGTH_INDEFINITE)
+    Snackbar.make(this, message, Snackbar.LENGTH_INDEFINITE)
         .setAnchorView(view)
         .also { snackbar ->
             snackbar.setAction("OK") {
@@ -96,36 +71,13 @@ fun View.snackbar(message: String, view: View) {
 }
 
 fun View.snackbar(message: String) {
-    Snackbar
-        .make(this, message, Snackbar.LENGTH_INDEFINITE)
+    Snackbar.make(this, message, Snackbar.LENGTH_INDEFINITE)
         .also { snackbar ->
             snackbar.setAction("OK") {
                 snackbar.dismiss()
             }
                 .show()
         }
-}
-
-fun AuthSignUpOptions.Builder<*>.defaultBuilder(authData: AuthData): AuthSignUpOptions {
-    return userAttributes(
-        mutableListOf
-        (
-            AuthUserAttribute(AuthUserAttributeKey.email(), authData.email),
-            AuthUserAttribute(AuthUserAttributeKey.custom(AppConstant.AUTH_ATTRIBUTE_CUSTOM_PROFILE), authData.thumbnail),
-            AuthUserAttribute(AuthUserAttributeKey.name(), authData.fullname)
-        )
-    ).build()
-}
-
-fun AuthSignUpOptions.Builder<*>.defaultBuilder(user: User): AuthSignUpOptions {
-    return userAttributes(
-        mutableListOf
-        (
-            AuthUserAttribute(AuthUserAttributeKey.email(), user.email),
-            AuthUserAttribute(AuthUserAttributeKey.custom("custom:profile_image"), user.thumbnail),
-            AuthUserAttribute(AuthUserAttributeKey.name(), user.name)
-        )
-    ).build()
 }
 
 fun RelativeLayout.bottomNavHide() {
