+++ app/src/main/java/com/github/code/gambit/utility/sharedpreference/PreferenceManager.kt
@@ -1,4 +1,4 @@
-package com.github.code.gambit.utility
+package com.github.code.gambit.utility.sharedpreference
 
 import android.content.Context
 import android.content.SharedPreferences
